refactor: プログレスバーと食品リストをコンポーネント化（#8）
- flowbite-reactのカスタムテーマを`theme.js`に別ファイル化
