+++ frontend/src/theme/theme.js
@@ -1,0 +1,11 @@
+export const customTheme = {
+  progress: {
+    color: {
+      primaryLight: "bg-primary-light",
+      redBar: "bg-redBar"
+    },
+    size: {
+      lgPlus: "h-5",
+    }
+  },
+};
